Use build tags to split tests
